Relocate SimplePackage
